+++ packages/build-utils/src/fs/node-version.ts
@@ -1,10 +1,14 @@
+import { statSync } from 'fs';
 import { intersects, validRange } from 'semver';
 import { NodeVersion } from '../types';
 import { NowBuildError } from '../errors';
 import debug from '../debug';
 
+export type NodeVersionMajor = ReturnType<typeof getOptions>[number]['major'];
+
 function getOptions() {
   const options = [
+    { major: 20, range: '20.x', runtime: 'nodejs20.x' },
     { major: 18, range: '18.x', runtime: 'nodejs18.x' },
     {
       major: 16,
@@ -37,24 +41,47 @@ function getOptions() {
       discontinueDate: new Date('2020-01-06'),
     },
   ] as const;
-  if (process.env.VERCEL_ALLOW_NODEJS20 === '1') {
-    return [
-      { major: 20, range: '20.x', runtime: 'nodejs20.x' },
-      ...options,
-    ] as const;
-  }
   return options;
 }
 
-function getHint(isAuto = false) {
-  const { major, range } = getLatestNodeVersion();
+function isNodeVersionAvailable(version: NodeVersion): boolean {
+  try {
+    return statSync(`/node${version.major}`).isDirectory();
+  } catch {
+    // ENOENT, or any other error, we don't care about
+  }
+  return false;
+}
+
+export function getAvailableNodeVersions(): NodeVersionMajor[] {
+  return getOptions()
+    .filter(isNodeVersionAvailable)
+    .map(n => n.major);
+}
+
+function getHint(isAuto = false, availableVersions?: NodeVersionMajor[]) {
+  const { major, range } = getLatestNodeVersion(availableVersions);
   return isAuto
     ? `Please set Node.js Version to ${range} in your Project Settings to use Node.js ${major}.`
     : `Please set "engines": { "node": "${range}" } in your \`package.json\` file to use Node.js ${major}.`;
 }
 
-export function getLatestNodeVersion() {
-  return getOptions()[0];
+export function getLatestNodeVersion(availableVersions?: NodeVersionMajor[]) {
+  const all = getOptions();
+  if (availableVersions) {
+    // Return the first node version that is definitely
+    // available in the build-container.
+    for (const version of all) {
+      for (const major of availableVersions) {
+        if (version.major === major) {
+          return version;
+        }
+      }
+    }
+  }
+  // As a fallback for local `vc build` and the tests,
+  // return the first node version if none is found.
+  return all[0];
 }
 
 export function getDiscontinuedNodeVersions(): NodeVersion[] {
@@ -63,9 +90,10 @@ export function getDiscontinuedNodeVersions(): NodeVersion[] {
 
 export async function getSupportedNodeVersion(
   engineRange: string | undefined,
-  isAuto = false
+  isAuto = false,
+  availableVersions?: NodeVersionMajor[]
 ): Promise<NodeVersion> {
-  let selection: NodeVersion = getLatestNodeVersion();
+  let selection: NodeVersion | undefined;
 
   if (engineRange) {
     const found =
@@ -74,17 +102,27 @@ export async function getSupportedNodeVersion(
         // the array is already in order so return the first
         // match which will be the newest version of node
         selection = o;
-        return intersects(o.range, engineRange);
+        return (
+          intersects(o.range, engineRange) &&
+          (availableVersions?.length
+            ? availableVersions.includes(o.major)
+            : true)
+        );
       });
     if (!found) {
       throw new NowBuildError({
         code: 'BUILD_UTILS_NODE_VERSION_INVALID',
         link: 'http://vercel.link/node-version',
         message: `Found invalid Node.js Version: "${engineRange}". ${getHint(
-          isAuto
+          isAuto,
+          availableVersions
         )}`,
       });
     }
+  }
+
+  if (!selection) {
+    selection = getLatestNodeVersion(availableVersions);
   }
 
   if (isDiscontinued(selection)) {