+++ packages/static-build/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/static-build
 
+## 2.0.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.11
+
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.10
+
 ## 2.0.10
 
 ### Patch Changes